+++ src/xpn_client/xpni/xpni_ft/xpni_ft_write.c
@@ -22,8 +22,8 @@
 
         /* debugging */
         #if defined(XPNI_DEBUG)
-            printf("[%s:%d] xpni_ft_swrite(%d,%p,%d,%lu); \n",
-                   __FILE__,__LINE__,fd,buffer,(int)offset,(unsigned long)size);
+            printf("[%s:%d] xpni_ft_swrite(%d,%p,%lu,%lu); \n",
+                   __FILE__,__LINE__,fd,buffer,(unsigned long)offset,(unsigned long)size);
         #endif
 
         /* check params */
@@ -50,8 +50,8 @@
 			default:
                              ret = (-1);
 #if defined(XPNI_DEBUG)
-                             printf("[%s:%d] xpni_ft_swrite(%d,%p,%d,%lu): %d fail(s)\n",
-                                    __FILE__,__LINE__,fd,buffer,(int)offset,(unsigned long)size,fmeta_nerrors);
+                             printf("[%s:%d] xpni_ft_swrite(%d,%p,%lu,%lu): %d fail(s)\n",
+                                    __FILE__,__LINE__,fd,buffer,(unsigned long)offset,(unsigned long)size,fmeta_nerrors);
 #endif
 			     break;
 		}
@@ -73,8 +73,8 @@
 			default:
                              ret = (-1);
 #if defined(XPNI_DEBUG)
-                             printf("[%s:%d] xpni_ft_swrite(%d,%p,%d,%lu): %d fail(s)\n",
-                                    __FILE__,__LINE__,fd,buffer,(int)offset,(unsigned long)size,fmeta_nerrors);
+                             printf("[%s:%d] xpni_ft_swrite(%d,%p,%lu,%lu): %d fail(s)\n",
+                                    __FILE__,__LINE__,fd,buffer,(unsigned long)offset,(unsigned long)size,fmeta_nerrors);
 #endif
 			     break;
 		}
@@ -84,8 +84,8 @@
 	{
             ret = (-1);
 #if defined(XPNI_DEBUG)
-            printf("[%s:%d] xpni_ft_swrite(%d,%p,%d,%d): Unknow file system tag: '%s' (not %s or %s)\n",
-                   __FILE__,__LINE__,fd,buffer,(int)offset,size,
+            printf("[%s:%d] xpni_ft_swrite(%d,%p,%lu,%lu): Unknow file system tag: '%s' (not %s or %s)\n",
+                   __FILE__,__LINE__,fd,buffer,(unsigned long)offset,(unsigned long)size,
 		   fmeta_fsTag,FS_TAG_RAID5INNER,FS_TAG_RAID5OUTER);
 #endif
 	}
